Create simpy cobined with matplotlib
In this example, we run a simulation of 4 cars trying to park in a parking lot with a capacity of 2. We store the start and end times of each car's parking in the results list. After the simulation has finished, we extract the start and end times from the results and plot them using Matplotlib. The red dots represent the start times of each car's parking and the blue dots represent the end times.