Create graph work 1
In this code, we simulate the arrival and travel times for 50 cars using a list arrival_times and travel_times, respectively. The travel times are generated using a Gaussian distribution with a mean of AVG_TRAVEL_TIME and a standard deviation of 2.

Next, we calculate the total time needed for all cars to reach road b by iterating through the list of cars and keeping track of the maximum time required.

Finally, we plot a histogram of the travel times using matplotlib to visualize if there's a queueing problem.